Fixes PR feedback
